backend/requirements.txt

# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
geoalchemy2==0.14.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Requests
httpx==0.25.1
aiohttp==3.9.1
requests==2.31.0

# Data Processing & AI
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# Semantic Web & LOD
rdflib==7.0.0
json-ld==1.3.0

# FiWARE Integration
paho-mqtt==1.6.1

# Utilities
geojson==3.1.0
pyproj==3.6.1
python-dateutil==2.8.2

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality
black==23.12.0
ruff==0.1.8

# Development
jupyter==1.0.0
ipython==8.18.1
